\name{internal_functions}
\alias{msdci}
\alias{alpha_beta_iniz}
\alias{sigma2_err_iniz}
\alias{z_par_iniz}
\alias{variance_fc}
\alias{variance_fc_d}
\alias{variance_fc_di}
\alias{mean_fc}
\alias{alpha_fc}
\alias{beta_fc}
\alias{z_fc}
\alias{sigma2_err_fc}
\alias{logPost_MCMC_wcum}
\alias{cauchit_probs}
\alias{dim_range_prob}


\title{
Internal \code{multiMarker} functions
}
\description{
Internal functions not to be called by the user.
}

\keyword{internal}
